+++ QHyper/optimizers/grid_search.py
@@ -64,10 +64,11 @@ class GridSearch(Optimizer):
                 disable=self.disable_tqdm
             ))
         min_idx = np.argmin([result.value for result in results])
-        history = [[result.value for result in results]]
 
+        history = [OptimizationResult(res.value, params, [[res]])
+                   for res, params in zip(results, hyperparams)]
         return OptimizationResult(
             value=results[min_idx].value,
             params=hyperparams[min_idx],
-            history=history
+            history=[history]
         )